chore: use wlm ContainerPort instead of new ContainerPort
